Use screencap program instead of framebuffer
Mmapping framebuffer works for getting parameters, but fails at copying
real framebuffer. Fortunately screencap program can return full
framebuffer to its stdout and it should be portable to every Android
device.

So we create a pipe, do a fork, set pipe to stdout of child and execl
into screencap. This is repeated for every frame and... almost works.
There is still issue with colors. However the output is already usable

As can be seen in the description above, this is a hack. It has for sure
awful resource consumption, but in reality is quite fast (~1 fps?).