Add option for forcing rotation
When we base on framebuffer as a source of resolution, we get always
screen in vertical position. Unfortunately screencap tries to be smarter
and rotates screeshot when user rotates the phone. This is simple hack
to make horizontal image possible. Sideeffect is now that user has to
know if he wants horizontal or vertical at startup. Then after rotating
the phone to another position, image will break.